requirements.txt
fastapi
odmantic
passlib
python-jose
uvicorn
python-multipart